- ported the whole application back to using floats like pdfbox was in the first place. the performance impact really was neglible compared to the uglyness and maintainance hell that resulted.
- rewrote and _massively_ simplified the character spacing class. now it is just like it should have been in the first place :):)

+++ src/main/java/org/elacin/pdfextract/tree/LineNode.java
@@ -151,7 +151,7 @@ public class LineNode extends AbstractParentNode<WordNode, ParagraphNode> {
      * @return
      */
     public boolean isOnSameLine(final AbstractNode node) {
-        int otherMiddleY = node.getPosition().getY() + (node.getPosition().getHeight() / 2);
+        float otherMiddleY = node.getPosition().getY() + (node.getPosition().getHeight() / 2);
         return getPosition().getY() <= otherMiddleY && getPosition().getEndY() >= otherMiddleY;
     }
